Final tuning with token relay
