+++ backend/src/main/java/io/spring/cloud/samples/animalrescue/backend/KubernetesSecurityConfiguration.java
@@ -7,6 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.oauth2.server.resource.authentication.ReactiveJwtAuthenticationConverterAdapter;
 import org.springframework.security.web.server.SecurityWebFilterChain;
 
 @Configuration
@@ -17,7 +18,11 @@ public class KubernetesSecurityConfiguration {
 
 	@Bean
 	public SecurityWebFilterChain securityWebFilterChain(ServerHttpSecurity httpSecurity) {
-		return httpSecurity.httpBasic().disable().build();
+		httpSecurity.oauth2ResourceServer()
+		            .jwt()
+					.jwtAuthenticationConverter(new ReactiveJwtAuthenticationConverterAdapter(new UserNameJwtAuthenticationConverter()));
+
+		return httpSecurity.build();
 	}
 
 }
