+++ k8s/base/animal-rescue-backend.yaml
@@ -15,6 +15,8 @@ spec:
         - name: animal-rescue-backend
           image: springcloudservices/animal-rescue-backend
           env:
+            - name: spring.profiles.active
+              value: k8s
             - name: spring.security.oauth2.resourceserver.jwt.jwk-set-uri
               valueFrom:
                 secretKeyRef:
